Remove Hit leaderboard and hit response.
Removing this since it only computes hearts over the top 11 messages
for the period. Not all messages over the period. i.e. bad results.

To do this properly we need to crawl all the messages in the archive
and store the counts in durable storage, which we don't support
at the moment.